### 2026-01-18 #9845 @editorialbot
[PRE REVIEW]: qdiv: A Python package for microbial ecology analysis using the Hill number framework
**Submitting author:** @omvatten (<a href="http://orcid.org/0000-0002-9232-6096">Oskar Modin</a>)
**Repository:** https://github.com/omvatten/qdiv
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.0
