### 2026-02-17 #9845 @rich2355
Hi @erinyoung, @Ruslan-Kurmashev, @dr-joe-wirth, @gchure, @agerada, are you willing to review this submission for JOSS? Thanks!
