### 2026-02-19 #9845 @rich2355
Hi @kelly-sovacool, @Shaun-Jiang, are you willing to review this submission for JOSS? Thanks!
