### 2026-02-19 #9845 @kelly-sovacool
@rich2355 I'm unable to review at this time due to time constraints
