### 2026-03-02 #10089 @Ruslan-Kurmashev
**Good practices:** Licensing, documentation and releases are in place and the docs provide manual verification steps. However, the dependency specification allows pandas 3, but core ordination/db-RDA functionality can fail under pandas 3 Copy-on-Write because _pcoa_lingoes_ performs an in-place write to _DataFrame.values_ (_observed ValueError: underlying array is read-only_). This should be addressed either by avoiding in-place writes or by temporarily constraining pandas <3 until a fix is released. Adding minimal tests to cover these workflows would help prevent regressions.
