### 2026-03-02 #10089 @Ruslan-Kurmashev
**Functionality:** I confirmed several core functional claims using the bundled example dataset _(MicrobiomeData.load_example("Saheb-Alam2019_DADA2").rarefy())_:
– Hill-based alpha diversity (_diversity.naive_alpha, q=1_)
– Hill-based beta diversity (_diversity.naive_beta, q=1; 16×16 distance matrix with zero diagona_l)
– PERMANOVA (_stats.permanova_)
– Mantel test (_stats.mantel_)
– Hill-adapted Raup–Crick null model _(model.rcq_)
– Heatmap plotting (_plot.heatmap_), successfully producing _qdiv_heatmap.png_ and _qdiv_heatmap.pdf_.

However, I encountered a **reproducible error** that prevents confirming the ordination and db-RDA claims: both **plot.ordination** and **stats.dbrda** fail with _ValueError: underlying array is read-only in qdiv/stats/ordination_calculations.py::pcoa_lingoes_ at the line _np.fill_diagonal(D_df.values, 0.0)_. Under pandas 3 (Copy-on-Write always enabled), DataFrame.values may be a read-only NumPy view, so in-place modification triggers this exception.

**Environment where the failure occurs:** qdiv 4.0.0, pandas 3.0.1, numpy 2.4.2, Python 3.11.

**Reproduction:**
> import qdiv
from qdiv import diversity, stats, plot
obj = qdiv.MicrobiomeData.load_example("Saheb-Alam2019_DADA2").rarefy()
dis = diversity.naive_beta(obj, q=1)
plot.ordination(dis, obj.meta, color_by="location", savename="qdiv_ordination.png")
stats.dbrda(dis, obj.meta, by=["location","feed"], perm_n=99, perm_seed=42)
