### 2026-03-03 #10089 @Ruslan-Kurmashev
**Community guidelines:** I could not find clear guidance for third parties on contributing, reporting bugs or seeking support. The repository does not include **CONTRIBUTING/SUPPORT/Code of Conduct** files and the README do not describe how users should report problems or request help. Please add minimal community documentation and ideally GitHub issue templates for bug reports requests.
