### 2026-03-03 #10089 @Ruslan-Kurmashev
**State of the field:** The paper.md currently does not include a dedicated  _State of the field section_, which is required for the JOSS paper format. Please add this section.
