### 2026-03-03 #10089 @Ruslan-Kurmashev
@rich2355 yes still in process with this part "Software paper"
