### 2026-03-03 #10089 @Ruslan-Kurmashev
**Software Design:** The work includes a _“Software design”_ section, but it currently reads mostly as a brief code-structure overview. Please expand it to explicitly discuss key design and architecture choices, the trade-offs considered and why these choices matter for research workflows.
