### 2026-03-03 #10089 @Ruslan-Kurmashev
**AI usage disclosure:** The work includes an AI usage disclosure, but it should be more specific and aligned with JOSS policy. Please clarify where generative AI was used (which code components, files and which manuscript sections), the tool context (Copilot version, plan and approximate date), and add an explicit statement that all AI-assisted code was reviewed, modified as needed, tested by the authors, who take full responsibility for the content.
