### 2026-03-03 #10089 @Ruslan-Kurmashev
**Quality of writing:** The manuscript requires editing for structure and readability. In paper.md, headings, sections and paragraphs appear insufficiently formatted as standard Markdown which harms readability and may affect rendering. Please reformat the manuscript with clear section breaks and paragraph spacing and ensure all required sections are included and consistently written.
