### 2026-03-03 #10089 @erinyoung
I conducted a comprehensive functional verification of qdiv by attempting to reproduce the core workflows described in [the documentation](https://qdiv.readthedocs.io/). In summary, most functions appear to work as expected.

Notes from [Load and save data](https://qdiv.readthedocs.io/en/latest/tutorials/Load_save_data.html):
Using the provided example datasets (Saheb-Alam2019 and Modin2025), I verified that the MicrobiomeData class successfully ingests and synchronizes diverse bioinformatic formats, including DADA2 abundance tables, FASTA sequences, Newick trees, CoverM relative abundance files, and SingleM EBD outputs. The API's handling of data cleaning, works as expected. The internal validation logic effectively maintained index integrity across alpha and beta diversity calculations.

Notes from [Data management](https://qdiv.readthedocs.io/en/latest/tutorials/Subsetting.html):
I also followed the tutorials for Data Management, which involved the data manipulation and subsetting capabilities of qdiv using both the DADA2 and Deblur example datasets. The `sequences.consensus` function successfully identified and aligned 630 common features across different processing pipelines, providing a detailed summary of retained abundance percentages. I verified that `subset_samples` and `subset_taxa` correctly filtered the dataset while maintaining strict index synchronization across all data facets (metadata, taxonomy, and phylogeny).

Additionally, `subset_abundant`,  `merge_samples`, and `rename_features` all seemed to run as expected.

Notes from [Diversity](https://qdiv.readthedocs.io/en/latest/tutorials/Diversity.html):
I followed the tutorial for Diversity estimation. Using the `naive_alpha` and `phyl_alpha` functions, I confirmed that the package computes diversity across a continuum of orders. The results consistently showed the expected mathematical scaling, where higher values of q increase sensitivity to abundant taxa, thereby lowering the effective number of species. It also appears that `naive_beta` and `naive_multi_beta` correctly partitions diversity into alpha and gamma components. The multi-sample beta diversity analysis by metadata category (e.g., 'feed') provided intuitive dissimilarity indices from both local and regional viewpoints.

Notes from [Plot](https://qdiv.readthedocs.io/en/latest/tutorials/Plot.html):
I was able to create two of the example plots:

For the third plot, I ran into this error
```
Traceback (most recent call last):
  File "/workspaces/qdiv/test3.py", line 25, in <module>
    fig, ax, data1, data2 = qdiv.plot.ordination(dis, obj, color_by="location", shape_by="feed")
                            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/workspaces/qdiv/qdiv/plot/ordination_plots.py", line 398, in ordination
    payload = _extract_ordination_payload(ordination_results)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/workspaces/qdiv/qdiv/plot/ordination_plots.py", line 29, in _extract_ordination_payload
    res = pcoa_lingoes(ordination_results)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/workspaces/qdiv/qdiv/stats/ordination_calculations.py", line 199, in pcoa_lingoes
    np.fill_diagonal(D_df.values, 0.0)
  File "/usr/local/python/3.12.1/lib/python3.12/site-packages/numpy/lib/_index_tricks_impl.py", line 923, in fill_diagonal
    a.flat[:end:step] = val
    ~~~~~~^^^^^^^^^^^
ValueError: underlying array is read-only
```
I have created a corresponding issue on the qdiv repository at https://github.com/omvatten/qdiv/issues/1.

I verified two of the plotting functions by exporting them to static PNG files. While the output quality is high and appropriate for publication, I identified two critical technical issues in the ordination workflow that require Major Revision:

Read-Only Array Error: The `stats.pcoa_lingoes` function fails with a ValueError: underlying array is read-only because it attempts to modify the input distance matrix diagonal in-place. This is a compatibility issue with modern NumPy/Pandas and should be fixed by performing operations on a local copy within the function.

API Consistency: There appears to be a discrepancy between the tutorial documentation and the actual function signature for `plot.ordination`. The tutorial suggests passing the distance matrix directly, but the function requires pre-calculated coordinates from `stats.pcoa_lingoes`. The error handling for this mismatch is currently a generic TypeError.

Notes from [Null models](https://qdiv.readthedocs.io/en/latest/tutorials/Model.html):

I verified the performance and accuracy of the null model implementations ($RC_q$ and $NRI_q$). On the standard example dataset, the Raup-Crick ($RC_q$) model completed 100 iterations in 27.44 seconds, demonstrating efficient handling of the permutation logic. The Tree Distance Matrix generation and Net Relatedness Index ($NRI_q$) were notably fast, completing in 1.06s and 1.61s respectively. The results correctly produced Standardized Effect Sizes (ses) that allow for the statistical interpretation of community assembly processes. The use of tqdm progress bars provides essential feedback for these computationally intensive tasks.
