### 2026-03-03 #10089 @erinyoung
# My Review
The qdiv package provides a much-needed Pythonic implementation of the Hill number framework for microbial ecology. My functional testing confirmed that the software successfully handles diverse bioinformatic inputs (DADA2, CoverM, SingleM) and provides a robust MicrobiomeData class that maintains data integrity during complex manipulations. While the ecological math and visualization components are high-quality, I have identified a few Major Revisions required for the repository to meet JOSS's scholarly and technical standards.

## Major Revisions Required
- Missing Automated Tests: The repository currently lacks a /tests directory. JOSS requires a suite of automated tests (e.g., pytest) to ensure the mathematical accuracy of diversity calculations and long-term maintainability. There are already ample example files readily available for a number of tests.

- Fatal Error in Ordination: I encountered a ValueError: underlying array is read-only in stats.pcoa_lingoes (line 199). This appears to be a write-permissions issue with modern NumPy/Pandas views and must be fixed by working on a copy of the input matrix.

- Community Guidelines: The repository is missing a CONTRIBUTING.md file, which is required by JOSS to facilitate open-source collaboration.

- State of the Field: The paper.md should be updated to include an explicit comparison to related tools like scikit-bio (Python) and vegan (R) to clarify the "build vs. contribute" justification.

- References: This section is empty, but will likely get citations added to it as the 'State of the Field' section is improved upon.

## Minor Suggestions
- API Consistency: The plot.ordination function signature in the documentation/tutorials does not currently match the requirements of the code (it expects a results dictionary, but tutorials imply it takes a distance matrix).

- Documentation: I recommend adding a note regarding the accelerate extra (Numba) to highlight the potential speedups for null model simulations.
