### 2026-03-05 #10089 @Ruslan-Kurmashev
### Overall review summary.

Thank you for the submission. qdiv addresses a real need in microbial ecology by providing a Hill-number–based toolkit for diversity analysis and related community-ecology statistics in a Python workflow. The project shows sustained development history, provides an OSI-approved license, has documentation on ReadTheDocs and uses tagged releases.

**What I verified**
- Installation proceeds as documented in a clean conda environment and the documented smoke test using `MicrobiomeData.load_example(...); obj.info()` runs successfully.
- On the bundled example dataset I confirmed: Hill-based alpha/beta diversity, PERMANOVA, Mantel test, the Hill-adapted Raup–Crick null model (`rcq`) and heatmap plotting (successfully generated `qdiv_heatmap.png` and `qdiv_heatmap.pdf`).

### **Priority issues to address**

**Required-structure**

1. Missing required “State of the field” section

- `paper.md` currently lacks a dedicated `# State of the field` section required by the JOSS paper format. Please add it to compare qdiv to commonly used tools and provide a “build vs. contribute” justification, with appropriate software, package citations.

**Functionality regression (dependency compatibility)**

2. Ordination (PCoA) and db-RDA fail under pandas 3

- In my environment (qdiv 4.0.0, Python 3.11, pandas 3.0.1, numpy 2.4.2), both `plot.ordination` and `stats.dbrda` fail with `ValueError: underlying array is read-only` originating from `pcoa_lingoes` where the code writes in-place to `DataFrame.values`.
- Please fix by avoiding in-place writes to `DataFrame.values` (use a writable copy via `to_numpy(copy=True)` / `np.array(..., copy=True)`) and add a minimal regression test, CI check for ordination and db-RDA.

**Open-source practice**

3) Community guidelines are missing
- I could not find clear guidance for third parties on contributing, reporting issues or seeking support. Please add minimal community documentation (`CONTRIBUTING.md` + `SUPPORT.md` or equivalent README sections) and ideally issue templates.

_Additional recommended improvements_

- Software design: expand to explicitly discuss key architectural choices, trade-offs and why they matter for research workflows (beyond a brief code-structure overview).
- AI usage disclosure: make the disclosure more specific (where AI was used: which code, files and which manuscript sections) and add an explicit human verification, responsibility statement.
- Quality of writing: reformat `paper.md` with clear Markdown section breaks, paragraph spacing for readability and rendering.
- References: citation syntax looks correct, but key claims should be backed by verifiable citations (PyPI download stats should cite the source and use the correct 180-day window). Please also reformat `paper.bib` for readability (one entry per block) and ensure the new “State of the field” comparisons are properly cited.

I am happy to re-check after the authors address the items above.
