+++ qdiv/model/simulate.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-from typing import Optional, Union, Tuple
+from typing import Optional, Union, Tuple, Callable, Sequence, List, Dict
 
 # -----------------------------------------------------------------------------
 # Simulate community structure
@@ -271,3 +271,359 @@ def generate_interdependence_matrix(
         matrix = (matrix + matrix.T) / 2
     species_names = [f"{species_prefix}{i+1}" for i in range(n_species)]
     return pd.DataFrame(matrix, index=species_names, columns=species_names)
+
+
+# -----------------------------------------------------------------------------
+# Generate tree dataframe
+# -----------------------------------------------------------------------------
+def make_block_tree_df(
+    k_per_level: Sequence[int],
+    *,
+    branch_length: float | Sequence[float] | Callable[[int, str, int], float] = 1.0,
+    root_name: str = "Root",
+    leaf_prefix: str = "OTU",
+    internal_prefix: str = "in",
+) -> pd.DataFrame:
+    """
+    Generate a block tree where branching factor varies with depth.
+    Compatible with phylo_utils (nodes, parent, branchL, leaves, dist_to_root).
+
+    Parameters
+    ----------
+    k_per_level : sequence of int
+        k_per_level[level] = number of children created at this depth.
+        Length of k_per_level = total depth.
+    branch_length :
+        float                → same length everywhere
+        sequence[len=depth]  → branch_length[level]
+        callable(level, parent_name, child_index) → full control
+    """
+    depth = len(k_per_level)
+
+    # ---- Normalize BL rule --------------------------------------------------
+    if isinstance(branch_length, (int, float)):
+        def resolve_bl(level, parent, j):
+            return float(branch_length)
+
+    elif isinstance(branch_length, (list, tuple, np.ndarray)):
+        if len(branch_length) != depth:
+            raise ValueError("branch_length must have length equal to depth")
+        def resolve_bl(level, parent, j):
+            return float(branch_length[level])
+
+    elif callable(branch_length):
+        def resolve_bl(level, parent, j):
+            return float(branch_length(level, parent, j))
+
+    else:
+        raise TypeError("branch_length must be float, sequence, or callable")
+
+    # ---- Storage ------------------------------------------------------------
+    nodes = []
+    parents = []
+    branchL = []
+    dist = []
+    children_map = {}
+    dist_map = {root_name: 0.0}
+
+    internal_counter = 0
+    leaf_counter = 0
+
+    # queue entries: (node_name, level)
+    queue = [(root_name, 0)]
+    parents_map = {root_name: None}
+
+    # Track order to preserve reproducible child ordering
+    while queue:
+        name, level = queue.pop(0)
+
+        # Record this row
+        p = parents_map[name]
+        if p is None:
+            nodes.append(name)
+            parents.append(None)
+            branchL.append(0.0)
+            dist.append(0.0)
+        else:
+            # find j: index of this node among parent's children
+            j = children_map[p].index(name)
+            L = resolve_bl(level-1, p, j)
+            branchL.append(float(L))
+            d = dist_map[p] + float(L)
+            dist_map[name] = d
+            nodes.append(name)
+            parents.append(p)
+            dist.append(d)
+
+        # Expand children only if below final level
+        if level < depth:
+            k = k_per_level[level]
+            kids = []
+            for j in range(k):
+                if level == depth - 1:
+                    # next level would be leaves
+                    leaf_counter += 1
+                    child = f"{leaf_prefix}{leaf_counter}"
+                else:
+                    internal_counter += 1
[…]
