+++ qdiv/diversity/beta_div.py
@@ -1,11 +1,38 @@
 import pandas as pd
 import numpy as np
 import math
-from typing import Optional, Dict, Any, Union
-from tqdm import tqdm
+from typing import Optional, Dict, Any, Union, Tuple
 from ..io import subset_samples
-from ..utils import rao, beta2dist, get_df, subset_tree, parse_leaves
+from ..utils import rao, beta2dist, get_df
+from ..utils import subset_tree_df, ra_to_branches, compute_Tmean
 from .alpha_div import naive_alpha, phyl_alpha, func_alpha
+
+def _get_tqdm(use_tqdm: bool):
+    """
+    Internal helper that returns tqdm if available and requested; otherwise provides
+    a minimal stub compatible with tqdm's API.
+    """
+    if use_tqdm:
+        try:
+            from tqdm import tqdm
+            return tqdm
+        except Exception:
+            pass
+
+    class _DummyTqdm:  # fallback with same constructor signature
+        def __init__(self, iterable=None, total=None, desc=None, unit=None, leave=False):
+            self._iterable = iterable if iterable is not None else range(total or 0)
+
+        def __iter__(self):
+            return iter(self._iterable)
+
+        def update(self, *_args, **_kwargs):
+            return None
+
+        def close(self):
+            return None
+
+    return _DummyTqdm
 
 # -----------------------------------------------------------------------------
 # Naive beta diversity
@@ -225,32 +252,9 @@ def phyl_beta(
         raise ValueError("`tab` must contain ≥ 2 samples (columns).")
 
     #Subset tree to features in tab
-    tree = subset_tree(tree, ra.index.tolist())
-
-    # Get Tmean
-    Tmean = tree[~tree['nodes'].str.startswith('in')].copy()
-    Tmean = Tmean.set_index('nodes')
-    asv_set = set(ra.index)
-    in_common = list(asv_set.intersection(Tmean.index.tolist()))
-    if len(in_common) < len(asv_set):
-        raise ValueError("Not all features in tab are found in the tree")
-    Tmean = Tmean.loc[in_common]
-    Tmean = Tmean['dist_to_root'].sum() / len(Tmean)
-
-    # --- Build branch × sample abundance matrix ------------------------------
-    # Initialize with zeros; ensure branch index matches tree
-    tree2 = pd.DataFrame(0.0, index=tree.index, columns=ra.columns)
-
-    for branch in tree.index:
-        leaves = parse_leaves(tree.loc[branch, "leaves"])
-
-        # Keep only features present in the abundance table
-        leaves = list(asv_set.intersection(leaves))
-        if leaves:
-            # Sum relative abundances for all leaves under the branch
-            tree2.loc[branch] = ra.loc[leaves].sum(axis=0)
-        else:
-            tree2.loc[branch] = 0.0
+    tree = subset_tree_df(tree, ra.index.tolist())
+    Tmean = compute_Tmean(tree, ra.index.tolist())
+    tree2 = ra_to_branches(ra, tree)
 
     # Align branch lengths to tree2 index
     branchL = tree["branchL"].reindex(tree2.index)
@@ -323,6 +327,7 @@ def func_beta(
     dis: bool = True,
     viewpoint: str = "regional",
     use_values_in_tab: bool = False,
+    use_tqdm: bool = True,
 ) -> pd.DataFrame:
     """
     Compute functional pairwise beta diversity of order *q*.
@@ -357,6 +362,8 @@ def func_beta(
     use_values_in_tab : bool, default=False
[…]
