+++ qdiv/utils/__init__.py
@@ -1,3 +1,5 @@
-from .phylo_utils import parse_newick, subset_tree, to_newick, parse_leaves, rename_leaves
-from .data_utils import beta2dist, rao, sort_index_by_number, get_df, rename_features
+from .phylo_utils import parse_newick, tree_to_dataframe, dataframe_to_tree
+from .phylo_utils import subset_tree, tree_to_newick, reroot_midpoint, ladderize_tree_df
+from .phylo_utils import parse_leaves, rename_leaves, subset_tree_df, ra_to_branches, compute_Tmean
+from .data_utils import beta2dist, rao, sort_index_by_number, get_df, rename_features, clean_taxonomy_table
 from .plot_utils import get_colors_markers, groupbytaxa

[…]
