+++ qdiv/diversity/beta_div.py
@@ -1491,329 +1491,9 @@ def dissimilarity_by_feature(
     
     return out
 
-
 # -----------------------------------------------------------------------------
-# Partition beta diversity
+# beta MPDq and MNTDq
 # -----------------------------------------------------------------------------
-
-# Structure-only dissimilarity on rank-sorted abundances (identity removed)
-def _get_structural_dis(
-    ra: pd.DataFrame,
-    *,
-    q: float = 1.0,
-) -> pd.DataFrame:
-    """
-    Compute structure-only Hill-number dissimilarity by:
-      1) building relative abundances from obj['tab'] (unless already relative)
-      2) sorting each sample's abundance vector descending (erases identity)
-      3) computing the same Hill-number dissimilarity on the sorted table
-
-    Parameters
-    ----------
-    ra : DataFrame
-        Relative abundance table.
-    q : float
-        Diversity order.
-
-    Returns
-    -------
-    dis_sorted : DataFrame (samples × samples)
-        Structure-only dissimilarity (identity erased).
-    """
-    # Vectorized column-wise sort (ascending), then flip to descending
-    a = np.sort(ra.to_numpy(dtype=float), axis=0)[::-1, :]
-    ra_sorted = pd.DataFrame(a, index=range(ra.shape[0]), columns=ra.columns)
-
-    # Compute structure-only dissimilarity using the same taxonomic Hill metric
-    dis_sorted = naive_beta(ra_sorted, q=q, dis=True, viewpoint="local", use_values_in_tab=True)
-
-    # Sanity: ensure sample order matches original
-    if not (dis_sorted.index.equals(ra.columns) and dis_sorted.columns.equals(ra.columns)):
-        dis_sorted = dis_sorted.reindex(index=ra.columns, columns=ra.columns)
-
-    return dis_sorted
-
-# A q-weighted mass-based approach get weights
-def _get_unique_fractions(
-    ra: pd.DataFrame,
-    *,
-    q: float = 1.0,
-) -> Tuple[float, float]:
-    """
-    Compute attribution weights u_A, u_B for two samples based on
-    community-specific 'unique' fractions at Hill order q, using a
-    min-based, abundance-compatible directional overlap.
-
-    For two columns A,B in `ra` (features × 2), define
-      C^{A|B}_{q,min} = sum_i min(p_iA, p_iB)^q / sum_i p_iA^q
-      C^{B|A}_{q,min} = sum_i min(p_iA, p_iB)^q / sum_i p_iB^q
-
-    Then the local unique fractions are
-      U_A^loc = 1 - C^{A|B}_{q,min}
-      U_B^loc = 1 - C^{B|A}_{q,min}
-
-    and weights are
-      λ_A = U_A^loc / (U_A^loc + U_B^loc),  λ_B = 1 - λ_A
-    with a 0.5/0.5 tie-break when both uniques are 0.
-
-    Parameters
-    ----------
-    ra : DataFrame
-        Relative abundances, shape (features × 2). Columns are the two samples.
-        Columns should sum to 1 (or will be normalized if eps>0 and sums>0).
-    q : float, default=1.0
-        Hill order.
-
-    Returns
-    -------
-    (w1, w2) : Tuple[float, float]
-        Weights for the first and second sample (λ_A, λ_B), in [0,1] and summing to 1.
-    """
-
-    def _get_qvector(vec, q):
-        vec = np.asarray(vec, dtype=float)
-        if q == 0:
[…]
