+++ qdiv/diversity/__init__.py
@@ -7,7 +7,7 @@ Provides functions for alpha and beta diversity calculations.
 
 from .alpha_div import naive_alpha, phyl_alpha, func_alpha, mpdq, mntdq
 from .beta_div import naive_beta, phyl_beta, func_beta, bray, jaccard, naive_multi_beta, phyl_multi_beta, func_multi_beta, evenness, dissimilarity_by_feature
-from .beta_div import partition_naive_beta, partition_phyl_beta, beta_mpdq, beta_mntdq
+from .beta_div import beta_mpdq, beta_mntdq
 
 __all__ = [
     "naive_alpha",
@@ -25,8 +25,6 @@ __all__ = [
     "func_multi_beta",
     "evenness",
     "dissimilarity_by_feature",
-    "partition_naive_beta",
-    "partition_phyl_beta",
     "beta_mpdq",
     "beta_mntdq",
 ]
