+++ SUPPORT.md
@@ -1,0 +1,22 @@
+# Support for qdiv
+This page provides  guidance for users who need help with qdiv.
+
+
+## Before Requesting Support
+Before opening a support request, please:
+
+- Review the README and documentations.
+- Search existing Issues to see whether your question has already been answered
+- Check CONTRIBUTING.md if your question relates to contributing code or reporting bugs
+
+
+## How to Get Help
+If you still need assistance, the recommended way to get support is to open a GitHub issue.
+Please include:
+
+- A clear description of the problem or question
+- qdiv version
+- Python version
+- Operating system
+- Minimal reproducible example (if applicable)
+
