+++ CONTRIBUTING.md
@@ -1,0 +1,35 @@
+# Contributing to qdiv
+Thank you for your interest in contributing to qdiv! Bug reports, feature requests, code contributions, and improvements to documentation are welcome!
+
+
+## Reporting Bugs
+If you encounter a bug, please open a GitHub issue.
+Include the following information:
+
+- qdiv version
+- Python version
+- Operating system
+- A minimal reproducible example
+- Expected vs. actual behaviour
+
+
+## Requesting Features
+To request a new feature. Check existing issues to avoid duplicates.
+Describe the problem or need and the proposed solution.
+
+
+## Contributing Code
+Fork the repository. Clone your fork and create a new branch.
+```bash
+git checkout -b feature/my-new-feature
+```
+
+Make your changes and submit a Pull Request describing your changes.
+
+
+## Improving documentation
+To contribute to the documentation:
+
+- Edit the relevant documentation file
+- Open a Pull Request summarizing your updates
+
