+++ docs/source/qdiv.stats.data_stats.rst
@@ -1,0 +1,7 @@
+qdiv.stats.data\_stats module
+=================================
+
+.. automodule:: qdiv.stats.data_stats
+   :members:
+   :show-inheritance:
+   :undoc-members:
