+++ docs/source/qdiv.stats.rst
@@ -10,6 +10,9 @@ qdiv.stats
    qdiv.stats.dbrda
    qdiv.stats.marginal_factor_tests_dbrda
    qdiv.stats.summarize_dbrda
+   qdiv.stats.corr
+   qdiv.stats.bootstrap_sample_matrix
+
 
 .. automodule:: qdiv.stats
    :members:
