+++ qdiv/stats/data_stats.py
@@ -10,9 +10,8 @@ from itertools import combinations, product
 import warnings
 from typing import List, Optional, Dict, Union, Any, Tuple
 from ..utils import get_df
-from .distance_tests import mantel
 
-__all__ = ["corr", "bootstrap_sample_matrix", "phylo_signal_mantel"]
+__all__ = ["corr", "bootstrap_sample_matrix"]
 
 # ------Helpers for corr ---------------------------
 def _is_categorical(series: pd.Series, max_unique: int = 20) -> bool:
@@ -688,508 +687,3 @@ def bootstrap_sample_matrix(
             },
         }
     return out
-
-
-# -----------------------------------------------------------------------------
-# Bootstrap function for confidence intervals
-# -----------------------------------------------------------------------------
-def phylo_signal_mantel(
-    obj: Union[Dict[str, Any], Any],
-    distmat: pd.DataFrame,
-    by: str,
-    *,
-    method: str = "pearson",
-    permutations: int = 999,
-    relative_abundance: bool = True,
-    min_total: float = 0.0,
-    verbose: bool = False,
-) -> Dict[str, Any]:
-    """
-    Estimate phylogenetic signal of species environmental niches using Mantel test.
-
-    Steps:
-        1. Extract abundance table (tab), metadata (meta), and phylo distmat.
-        2. Extract environmental variable from meta (by column).
-        3. Compute species-level niche value as abundance-weighted mean environment:
-               niche_i = sum_s(abundance_{i,s} * env_s) / sum_s(abundance_{i,s})
-        4. Compute pairwise absolute-niche-distance matrix between species.
-        5. Mantel test between species-niche-distance and phylogenetic distance matrix.
-    
-    Parameters
-    ----------
-    obj : MicrobiomeData or dict
-        Must contain 'tab' and 'meta'.
-    distmat : DataFrame
-        Square phylogenetic distance matrix (features × features).
-    by : str
-        Column name in metadata to extract environmental values (e.g., "pH").
-    method : {"pearson","spearman"}, optional
-        Correlation method for Mantel.
-    permutations : int, optional
-        Number of null permutations.
-
-    Returns
-    -------
-    dict
-        {
-            "by"        : environmental variable name,
-            "niche"      : species niche vector (aligned with distmat),
-            "niche_dist_vec": distance vector that could be used for plotting
-            "phylo_dist_vec": distance vector that could be used for plotting
-            "mantel_r"   : Mantel correlation,
-            "mantel_p"   : p-value,
-            "mantel_out" : full mantel output
-        }
-    """
-    # --- extract data ---
-    tab = get_df(obj, "tab")
-    meta = get_df(obj, "meta")
-
-    if tab is None or meta is None:
-        raise ValueError("Need 'tab' and 'meta' in object.")
-
-    if by not in meta.columns:
-        raise ValueError(f"Environmental variable '{by}' not found in metadata.")
-
-    # Align metadata values to tab sample order
-    env_raw = meta.loc[tab.columns, by]
-    
-    # ----- abundances -----
-    A = tab.copy().astype(float)
-    A = A.div(A.sum(axis=0), axis=1).fillna(0.0)
-
[…]
