+++ qdiv/stats/__init__.py
@@ -7,7 +7,7 @@ Provides functions for multivariate statistical tests and ordinations.
 
 from .distance_tests import mantel, permanova, gower
 from .ordination_calculations import pcoa_lingoes, dbrda, summarize_dbrda
-from .data_stats import corr, bootstrap_sample_matrix, phylo_signal_mantel
+from .data_stats import corr, bootstrap_sample_matrix
 
 __all__ = [
     "mantel",
@@ -18,5 +18,4 @@ __all__ = [
     "summarize_dbrda",
     "corr",
     "bootstrap_sample_matrix",
-    "phylo_signal_mantel"
 ]
