+++ tests/test_null_models_placeholder.py
@@ -1,0 +1,8 @@
+
+import pytest
+
+@pytest.mark.slow
+def test_null_models_placeholder_documentation():
+    # This is a placeholder for heavier null-model tests (rcq, nriq).
+    # Intentionally empty to keep CI fast; enable real tests locally if desired.
+    assert True
