+++ pytest.ini
@@ -1,0 +1,11 @@
+
+[pytest]
+minversion = 7.0
+addopts = -ra
+filterwarnings =
+    ignore:.*is a deprecated alias for.*:DeprecationWarning
+    ignore:.*distutils Version classes are deprecated.*:DeprecationWarning
+    ignore:.*setdefaultencoding.*:DeprecationWarning
+    ignore:.*Copy-on-Write.*:FutureWarning
+markers =
+    slow: marks tests as slow (deselect with '-m "not slow"')
