+++ .github/workflows/ci.yml
@@ -1,0 +1,33 @@
+
+name: CI (pytest)
+
+on:
+  push:
+    branches: [ main, master ]
+  pull_request:
+    branches: [ main, master ]
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+        pandas: ["2.2.*", "3.*"]
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+          pip install "pandas==${{ matrix.pandas }}" pytest matplotlib
+      - name: Run pytest
+        env:
+          MPLBACKEND: Agg
+        run: |
+          pytest -q
