+++ .github/workflows/ci.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
+        python-version: ["3.11", "3.12"]
         pandas: ["2.2.*", "3.*"]
     steps:
       - uses: actions/checkout@v4
