+++ qdiv/stats/distance_tests.py
@@ -18,6 +18,7 @@ from pandas.api.types import (
     is_datetime64_any_dtype,
     is_bool_dtype,
     is_categorical_dtype,
+    CategoricalDtype,
 )
 
 __all__ = [
@@ -184,106 +185,6 @@ def mantel(
     p = (np.sum(null_stats <= obs) + 1) / (permutations + 1)
     return [obs, p]
 
-
-# def mantel(
-#     dis1: pd.DataFrame,
-#     dis2: pd.DataFrame,
-#     method: Literal["spearman", "pearson", "absDist"] = "spearman",
-#     getOnlyStat: bool = False,
-#     permutations: int = 999
-# ) -> Union[float, List[float]]:
-#     """
-#     Perform a Mantel test between two dissimilarity matrices.
-
-#     The Mantel test evaluates the association between two distance/dissimilarity
-#     matrices by comparing their lower‑triangular entries. A permutation test
-#     is used to assess significance by randomly permuting sample labels in one
-#     matrix.
-
-#     Parameters
-#     ----------
-#     dis1 : pandas.DataFrame
-#         First square dissimilarity matrix (samples × samples).
-#     dis2 : pandas.DataFrame
-#         Second square dissimilarity matrix (samples × samples).
-#     method : {'spearman', 'pearson', 'absDist'}, default='spearman'
-#         Correlation/dissimilarity measure:
-#         - 'spearman' : Spearman rank correlation (returns 1 − ρ)
-#         - 'pearson'  : Pearson correlation (returns 1 − r)
-#         - 'absDist'  : Mean absolute difference between distances
-#     getOnlyStat : bool, default=False
-#         If True, return only the observed statistic (no permutations).
-#     permutations : int, default=999
-#         Number of permutations for the null distribution.
-
-#     Returns
-#     -------
-#     float or list [statistic, p_value]
-#         - If getOnlyStat=True: returns the observed statistic.
-#         - Otherwise: returns [observed_statistic, p_value].
-
-#     Notes
-#     -----
-#     - Matrices are automatically reordered to have identical sample order.
-#     - Only the lower triangular part (excluding diagonal) is used.
-#     - For correlation methods, the statistic is expressed as a *dissimilarity*
-#       (1 − r or 1 − ρ), so **smaller values indicate stronger similarity**.
-#     """
-
-#     # --- Input validation ----------------------------------------------------
-#     if not isinstance(dis1, pd.DataFrame) or not isinstance(dis2, pd.DataFrame):
-#         raise TypeError("dis1 and dis2 must be pandas DataFrames.")
-
-#     if dis1.shape != dis2.shape:
-#         raise ValueError("dis1 and dis2 must have the same shape.")
-
-#     if method not in {"spearman", "pearson", "absDist"}:
-#         raise ValueError("method must be 'spearman', 'pearson', or 'absDist'.")
-
-#     # --- Align matrices by sorted sample names -------------------------------
-#     samples = sorted(dis1.columns.tolist())
-#     dis1 = dis1.loc[samples, samples]
-#     dis2 = dis2.loc[samples, samples]
-
-#     # --- Helper: compute Mantel statistic ------------------------------------
-#     def get_stat(mat1: pd.DataFrame, mat2: pd.DataFrame) -> float:
-#         mask = np.tril(np.ones(mat1.shape, dtype=bool), k=-1)
-
-#         v1 = mat1.values[mask]
-#         v2 = mat2.values[mask]
-
-#         if method in ("spearman", "pearson"):
-#             dfcorr = pd.DataFrame({"x": v1, "y": v2})
-#             corr = dfcorr.corr(method=method).loc["x", "y"]
-#             return 1 - corr  # convert similarity → dissimilarity
-
-#         elif method == "absDist":
[…]
