+++ tests/test_null_models.py
@@ -7,39 +7,18 @@ def test_rcq_runs_and_shapes(qb):
     Smoke + structure test for the Hill-adapted Raup–Crick null model (RC_q).
     Uses a tiny iteration count so CI remains fast.
     """
-    dis = qdiv.diversity.naive_beta(qb, q=1)
-
     # run the rcq model with few iterations
-    out = qdiv.model.rcq(dis, qb.meta, by="feed", q=1, n_iter=12, seed=7)
+    out = qdiv.model.rcq(qb, q=1, iterations=12, random_state=7)
 
-    # out should be a DataFrame
-    assert isinstance(out, pd.DataFrame)
+    # out should be a Dict
+    assert isinstance(out, dict)
 
     # output must index samples
-    assert list(out.index) == list(qb.meta.index)
+    assert list(out["p"].index) == list(qb.meta.index)
 
     # SES column must be present
-    assert "ses" in out.columns
+    assert "ses" in out.keys()
 
-    # ses values must be numeric or NaN
-    ses = out["ses"].astype(float)
-    assert np.all(np.isfinite(ses) | np.isnan(ses))
-
-def test_nriq_runs_and_shapes(qb):
-    """
-    Smoke + structure test for NRI_q (Net Relatedness Index with Hill weights).
-    """
-    # NRI uses phylogeny; ensure obj has a tree (example datasets do)
-    out = qdiv.model.nriq(qb, q=1, n_iter=12, seed=11)
-
-    # output format: DataFrame indexed by sample
-    assert isinstance(out, pd.DataFrame)
-    assert list(out.index) == list(qb.meta.index)
-
-    # should contain at least the SES (standardized effect size)
-    assert "ses" in out.columns
-    ses = out["ses"].astype(float)
-    assert np.all(np.isfinite(ses) | np.isnan(ses))
 
 def test_null_models_repeatability(qb):
     """
@@ -47,12 +26,12 @@ def test_null_models_repeatability(qb):
     Same seed => same output.
     Different seed => different output often.
     """
-    out1 = qdiv.model.rcq(qb, qb.meta, by="feed", q=1, n_iter=12, seed=5)
-    out2 = qdiv.model.rcq(qb, qb.meta, by="feed", q=1, n_iter=12, seed=5)
+    out1 = qdiv.model.rcq(qb, q=1, iterations=12, random_state=5)
+    out2 = qdiv.model.rcq(qb, q=1, iterations=12, random_state=5)
 
     # Exactly equal because same seed
-    assert out1.equals(out2)
+    assert out1["ses"].equals(out2["ses"])
 
     # Different seed should usually produce differences
-    out3 = qdiv.model.rcq(qb, qb.meta, by="feed", q=1, n_iter=12, seed=99)
-    assert not out1.equals(out3)
+    out3 = qdiv.model.rcq(qb, q=1, iterations=12, random_state=99)
+    assert not out1["ses"].equals(out3["ses"])
