+++ tests/test_stats_light.py
@@ -18,4 +18,4 @@ def test_permanova_and_mantel_smoke(qb, dis):
     # Mantel against itself should be near-perfect correlation
     mantel = qdiv.stats.mantel(dis, dis, permutations=59, random_state=7)
     r = float(mantel[0]) if isinstance(mantel, list) else float(mantel.get('r', mantel.get('statistic', 0.0)))
-    assert 0.0 <= r <= 0.05
+    assert -1e-12 <= r <= 0.05
