### 2026-03-09 #10089 @omvatten
Again, many thanks for the reviews! I’ve incorporated several of the suggestions and wanted to give a brief update:
1.	PCoA/dbRDA fix: The failure in stats.pcoa_lingoes has been resolved by working on a writable copy of the distance matrix. (Thanks @erinyoung for opening the issue.) The new release v4.0.1 includes this fix.
2.	Ordination API clarification: The documentation for plot.ordination has been updated. It now states that the function accepts either (i) a dissimilarity matrix (in which case pcoa_lingoes is run automatically), or (ii) result dictionaries returned by pcoa_lingoes or dbrda.
3.	Community guidelines: I’ve added CONTRIBUTING.md, SUPPORT.md, and CODE_OF_CONDUCT.md, and updated the README.
4.	Automated tests: A basic test suite has been added covering diversity calculations, PERMANOVA, Mantel tests, and one of the null models.
I’m currently working on improving the paper sections (State of the Field, Software Design, writing quality, etc.). I’ll post another update here once those revisions are completed.
