+++ paper.bib
@@ -1,4 +1,4 @@
-@article{Chao:2014,
+@article{chao2014,
   author = {{Chao}, A. and {Chiu}, C.-H. and {Jost}, L.},
   title = {Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers},
   journal = {Annual Review of Ecology, Evolution, and Systematics},
@@ -8,7 +8,51 @@
   url = {https://www.annualreviews.org/content/journals/10.1146/annurev-ecolsys-120213-091540}
 }
 
-@article{Raup:1979,
+@Article{hill1973,
+  author  = {Hill, Mark O.},
+  title   = {Diversity and evenness: A unifying notation and its consequences},
+  journal = {Ecology},
+  year    = {1973},
+  volume  = {54},
+  number  = {2},
+  pages   = {427--432},
+  doi     = {10.2307/1934352}
+}
+
+@Article{jost2007,
+  author  = {Jost, Lou},
+  title   = {Partitioning diversity into independent alpha and beta components},
+  journal = {Ecology},
+  year    = {2007},
+  volume  = {88},
+  number  = {10},
+  pages   = {2427--2439},
+  doi     = {10.1890/06-1736.1}
+}
+
+@Article{chiu2014,
+  author  = {Chiu, Chun-Huo and Jost, Lou and Chao, Anne},
+  title   = {Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers},
+  journal = {Ecological Monographs},
+  year    = {2014},
+  volume  = {84},
+  number  = {1},
+  pages   = {21--44},
+  doi     = {10.1890/12-0960.1}
+}
+
+@Article{chiuchao2014,
+  author  = {Chiu, Chun-Huo and Chao, Anne},
+  title   = {Distance-based functional diversity measures and their decomposition: a framework based on Hill numbers},
+  journal = {PLoS ONE},
+  year    = {2014},
+  volume  = {9},
+  number  = {7},
+  pages   = {e100014},
+  doi     = {10.1371/journal.pone.0100014}
+}
+
+@article{raup1979,
   author = {{Raup}, D.M. and {Crick}, R.E.},
   title = {Measurement of faunal similarity in paleontology},
   journal = {Journal of Paleontology},
@@ -17,7 +61,7 @@
   url = {https://www.jstor.org/stable/1304099}
 }
 
-@article{Webb:2002,
+@article{webb2002,
   author = {{Webb}, C.O. and {Ackerly}, D.D. and {McPeek}, M.A. and {Donoghue}, M.J.},
   title = {Phylogenies and Community Ecology},
   journal = {Annual Review of Ecology, Evolution, and Systematics},
@@ -27,7 +71,106 @@
   url = {https://www.annualreviews.org/content/journals/10.1146/annurev.ecolsys.33.010802.150448}
 }
 
-@article{Modin:2020,
+@Manual{vegan2026,
+  title = {vegan: Community Ecology Package},
[…]
