+++ paper.md
@@ -11,7 +11,7 @@ authors:
 affiliations:
  - name: Chalmers University of Technology, Department of Architecture and Civil Engineering, Division of Water Environment Technology, Gothenburg, Sweden
    index: 1
-date: 13 March 2026
+date: 15 March 2026
 bibliography: paper.bib
 ---
 
@@ -34,12 +34,12 @@ offering a streamlined and versatile tool for investigating patterns of communit
 
 `qdiv` is a Python package designed for microbial ecology. 
 It was developed to facilitate the management and analysis of data generated from amplicon sequencing and metagenomics within a Python programming environment. 
-Specifically, qdiv implements the Hill number framework [@chao2014] for assessing diversity across a broad range of ecological metrics and analyses. 
+Specifically, `qdiv` implements the Hill number framework [@chao2014] for assessing diversity across a broad range of ecological metrics and analyses. 
 This framework enables systematic evaluation of how relative abundances influence diversity measures.
 
 Traditionally, Hill numbers [@hill1973] are applied to quantify alpha diversity, which is the diversity of taxa within a single community. 
 However, the framework can also be applied to beta diversity [@jost2007], which measures differences in composition between communities, 
-and extended to metrics that incorporate phylogenetic and functional relationships among taxa [@chiu2014][@chiuchao2014].
+and extended to metrics that incorporate phylogenetic and functional relationships among taxa [@chiu2014; @chiuchao2014].
 
 Many ecological statistical tests and null models, such as Raup-Crick [@raup1979], the Net Relatedness Index, and the Nearest-Taxon Index [@webb2002], 
 have used presence-absence data or relied on a single relative abundance-based index (e.g. Bray-Curtis). 
@@ -49,9 +49,8 @@ have used presence-absence data or relied on a single relative abundance-based i
 # Statement of the field
 
 A variety of software tools support microbial community analysis, particularly within the R programming environment.
-Packages such as vegan [vegan2026], phyloseq [@mcmurdie2013], microbiome [@lahti207], and ampvis2 [@andersen2018] provide comprehensive workflows 
-for diversity calculations, ordination, and phylogenetic analysis, 
-but they primarily rely on classical indices rather than a unified diversity framework. 
+Packages such as vegan [vegan2026], phyloseq [@mcmurdie2013], microbiome [@lahti2017], and ampvis2 [@andersen2018] provide comprehensive workflows 
+for diversity calculations, ordination, and phylogenetic analysis, but they primarily rely on classical indices rather than a unified diversity framework. 
 Several R packages implement the Hill number framework, including hilldiv [@alberdi2019], hillR [@li2018], vegetarian [@vegetarian2019], 
 and iNEXT [@inext2024]. However, to our knowledge there is no Python package that implements the Hill number framework in a unified way 
 for diversity metrics and statistical workflows. The `qdiv` package fills this gap by offering a systematic Python implementation of 
@@ -60,15 +59,7 @@ Hill‑based alpha and beta diversity, phylogenetic diversity, and associated st
 Other commonly used tools include QIIME [@bolyen2019], an end‑to‑end microbiome analysis platform typically 
 used through a command‑line interface, which limits flexible, script‑based workflows in Python.
 The Python library scikit‑bio [@aton2025] provides data structures and algorithms for omics analyses, 
-including sequence handling, phylogenetic tree operations, and ecological statistics, 
-but it does not center around, or generalize, the Hill number framework.
-
-
-#Build vs. contribute justification
[…]
