+++ paper.bib
@@ -8,48 +8,48 @@
   url = {https://www.annualreviews.org/content/journals/10.1146/annurev-ecolsys-120213-091540}
 }
 
-@Article{hill1973,
-  author  = {Hill, Mark O.},
-  title   = {Diversity and evenness: A unifying notation and its consequences},
+@article{hill1973,
+  author = {Hill, Mark O.},
+  title = {Diversity and evenness: A unifying notation and its consequences},
   journal = {Ecology},
-  year    = {1973},
-  volume  = {54},
-  number  = {2},
-  pages   = {427--432},
-  doi     = {10.2307/1934352}
+  year = {1973},
+  volume = {54},
+  number = {2},
+  pages = {427--432},
+  doi = {10.2307/1934352}
 }
 
-@Article{jost2007,
-  author  = {Jost, Lou},
-  title   = {Partitioning diversity into independent alpha and beta components},
+@article{jost2007,
+  author = {Jost, Lou},
+  title = {Partitioning diversity into independent alpha and beta components},
   journal = {Ecology},
-  year    = {2007},
-  volume  = {88},
-  number  = {10},
-  pages   = {2427--2439},
-  doi     = {10.1890/06-1736.1}
+  year = {2007},
+  volume = {88},
+  number = {10},
+  pages = {2427--2439},
+  doi = {10.1890/06-1736.1}
 }
 
-@Article{chiu2014,
-  author  = {Chiu, Chun-Huo and Jost, Lou and Chao, Anne},
-  title   = {Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers},
+@article{chiu2014,
+  author = {Chiu, Chun-Huo and Jost, Lou and Chao, Anne},
+  title = {Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers},
   journal = {Ecological Monographs},
-  year    = {2014},
-  volume  = {84},
-  number  = {1},
-  pages   = {21--44},
-  doi     = {10.1890/12-0960.1}
+  year = {2014},
+  volume = {84},
+  number = {1},
+  pages = {21--44},
+  doi = {10.1890/12-0960.1}
 }
 
-@Article{chiuchao2014,
-  author  = {Chiu, Chun-Huo and Chao, Anne},
-  title   = {Distance-based functional diversity measures and their decomposition: a framework based on Hill numbers},
+@article{chiuchao2014,
+  author = {Chiu, Chun-Huo and Chao, Anne},
+  title = {Distance-based functional diversity measures and their decomposition: a framework based on Hill numbers},
   journal = {PLoS ONE},
-  year    = {2014},
-  volume  = {9},
-  number  = {7},
-  pages   = {e100014},
-  doi     = {10.1371/journal.pone.0100014}
+  year = {2014},
+  volume = {9},
+  number = {7},
+  pages = {e100014},
+  doi = {10.1371/journal.pone.0100014}
 }
 
 @article{raup1979,
@@ -71,7 +71,7 @@
   url = {https://www.annualreviews.org/content/journals/10.1146/annurev.ecolsys.33.010802.150448}
 }
 
-@Manual{vegan2026,
+@manual{vegan2026,
   title = {vegan: Community Ecology Package},
[…]
