### 2026-03-15 #10089 @omvatten
Now I have also updated paper.md and the references in paper.bib. The main changes include:
- A new Statement of the field section that positions qdiv among related tools.
- An expanded Software design section that clarified design choices and trade-offs.
- A more detailed AI usage disclosure  section.
Below, I provide point-by-point responses to the comments from @erinyoung and @Ruslan-Kurmashev.
