### 2026-03-15 #10089 @omvatten
Responses to comments from @erinyoung
> ## Major Revisions Required
> * Missing Automated Tests: The repository currently lacks a /tests directory. JOSS requires a suite of automated tests (e.g., pytest) to ensure the mathematical accuracy of diversity calculations and long-term maintainability. There are already ample example files readily available for a number of tests.
Response: A test suite has been added. The /tests directory is included in the repository and uses pytest.
> * Fatal Error in Ordination: I encountered a ValueError: underlying array is read-only in stats.pcoa_lingoes (line 199). This appears to be a write-permissions issue with modern NumPy/Pandas views and must be fixed by working on a copy of the input matrix.
Response: This has been corrected in v4.0.1.
> * Community Guidelines: The repository is missing a CONTRIBUTING.md file, which is required by JOSS to facilitate open-source collaboration.
Response: CONTRIBUTING.md, SUPPORT.md, and CODE_OF_CONDUCT.md have been added.
> * State of the Field: The paper.md should be updated to include an explicit comparison to related tools like scikit-bio (Python) and vegan (R) to clarify the "build vs. contribute" justification.
Response: A state of the field section has been added to paper.md. It explains how qdiv is positioned in comparison to related tools.
> * References: This section is empty, but will likely get citations added to it as the 'State of the Field' section is improved upon.
Response: References are provided in the paper.bib file and the paper cites relevant related work in the new State of the field section.
> ## Minor Suggestions
> * API Consistency: The plot.ordination function signature in the documentation/tutorials does not currently match the requirements of the code (it expects a results dictionary, but tutorials imply it takes a distance matrix).
Response: It has been clarified in the documentation that it accepts either a dissimilarity matrix (dataframe) or a results dictionary returned by the ordination functions.
> * Documentation: I recommend adding a note regarding the accelerate extra (Numba) to highlight the potential speedups for null model simulations.
Response: Numba acceleration is currently only available for the sequences subpackage. The installation instructions have been updated with this information both in the README.md and the online documentation.
