### 2026-03-15 #10089 @omvatten
Responses to comments from @Ruslan-Kurmashev
>
> ### **Priority issues to address**
> **Required-structure**
>
> 1. Missing required “State of the field” section
>
> * `paper.md` currently lacks a dedicated `# State of the field` section required by the JOSS paper format. Please add it to compare qdiv to commonly used tools and provide a “build vs. contribute” justification, with appropriate software, package citations.
Response: A state of the field section has been added to paper.md. It includes comparisons with existing R and Python tools, and adds citations.
> **Functionality regression (dependency compatibility)**
> 2. Ordination (PCoA) and db-RDA fail under pandas 3
> * In my environment (qdiv 4.0.0, Python 3.11, pandas 3.0.1, numpy 2.4.2), both `plot.ordination` and `stats.dbrda` fail with `ValueError: underlying array is read-only` originating from `pcoa_lingoes` where the code writes in-place to `DataFrame.values`.
> * Please fix by avoiding in-place writes to `DataFrame.values` (use a writable copy via `to_numpy(copy=True)` / `np.array(..., copy=True)`) and add a minimal regression test, CI check for ordination and db-RDA.
Response: This has been fixed in version 4.0.1.
> **Open-source practice**
> 3. Community guidelines are missing
> * I could not find clear guidance for third parties on contributing, reporting issues or seeking support. Please add minimal community documentation (`CONTRIBUTING.md` + `SUPPORT.md` or equivalent README sections) and ideally issue templates.
Response: CONTRIBUTING.md, SUPPORT.md, and CODE_OF_CONDUCT.md have been added, and the README has been updated.

> _Additional recommended improvements_
> * Software design: expand to explicitly discuss key architectural choices, trade-offs and why they matter for research workflows (beyond a brief code-structure overview).
Response: This section has been improved in paper.md.
> * AI usage disclosure: make the disclosure more specific (where AI was used: which code, files and which manuscript sections) and add an explicit human verification, responsibility statement.
Response: This section has been improved and a human verification and responsibility statement has been added.
> * Quality of writing: reformat `paper.md` with clear Markdown section breaks, paragraph spacing for readability and rendering.
Response: This has been revised and the rendering has been verified in the PDF generated by Github Actions.
> * References: citation syntax looks correct, but key claims should be backed by verifiable citations (PyPI download stats should cite the source and use the correct 180-day window). Please also reformat `paper.bib` for readability (one entry per block) and ensure the new “State of the field” comparisons are properly cited.
Response: The PyPI download stats as been updated with exact dates, and excludes mirrors. The paper.bib file has been updated with the new references.
