### 2026-03-19 #10089 @Shaun-Jiang
## My review:
I reviewed the modifications @omvatten made on both package content and script bugs in response to two reviewers, and tested the function incorporated in this package including:
- Installation proceeded without issues;
- Test dataset from DADA2 and Deblur included in Saheb-Alam_2019  were used for validating the functions within the package including loading, subsetting, and rarifying.
- Calculation of alpha and beta diversity (both naive and phylogenetic diversity). For functional diversity, I built the distance matrix based on the phylogeentic tree inlcuded in DADA2 dataset. The functions seems to work as intended.
- Plotting including heat map, diversity profiles, and PCoA
- Hill-adapted Raup–Crick null model
- Statistical analysis functions including PERMANOVA and Mantel test.

The authors have addressed all previously identified major concerns, including the addition of automated tests, resolution of the pandas 3 compatibility issue, and improvements to documentation and manuscript structure. The software is functional, well-scoped, and fills a clear gap in the Python ecosystem for Hill-number–based microbial ecology analysis.

A few minor revisions remain, primarily related to modification to the current user manuals (https://qdiv.readthedocs.io/en/latest/index.html). Although functional alpha and beta diversity calculation has been included, a demonstration using the example dataset is recommended to make this package more user friendly, given the input is different from naive and phylogenetic diversity. Similarly, popular statistical functions such as PERMANOVA and Mantel test can also be included in the documentation. Addressing these will further improve the robustness and scholarly clarity of the submission.
