### 2026-03-25 #10089 @omvatten
I have revised and expanded the documentation to examples for three statistical functions provided by qdiv: mantel, permanova, and bootstrap_sample_matrix. The updated documentation is available here:
https://qdiv.readthedocs.io/en/latest/tutorials/Statistical_functions.html.

I have also added an example of how to calculate functional diversity, see here: https://qdiv.readthedocs.io/en/latest/tutorials/Diversity.html#Alpha-diversity

While revising the documentation, I also identified and corrected some inconsistencies and potential issues in the implementation of the statistical functions. In particular, the argument used to control random number generation was inconsistently named (random_state in some functions and seed in others). This has now been standardized to random_state, with seed retained as a backward-compatible alias.
The corrections and documentation improvements are included in a new release of qdiv (v4.0.2).

Thank you all for the constructive comments and suggestions.
