### 2026-04-01 #10089 @Ruslan-Kurmashev
@rich2355 Ending signal: I have completed this review round.

I re-checked the updated release (tag v4.0.2) in a clean environment (Python 3.11, pandas 3.0.2, numpy 2.4.4). The previously reported pandas 3 Copy-on-Write failure is resolved: `plot.ordination` (PCoA) and `stats.dbrda` run successfully on the bundled example dataset and produce ordination outputs. I also ran the repository test suite on v4.0.2 (`python -m pytest -q`) and all tests passed (10 passed).

The authors have addressed the major review items (community guidelines files added, automated tests added, paper sections updated). I consider the submission ready from my side.

Minor note (non-blocking): `import qdiv; qdiv.__version__` reports 4.0.0 even when the installed distribution is 4.0.2
