### 2026-04-07 #10089 @omvatten
Hi @rich2355! I have addressed @Shaun-Jiang 's comments in this post: https://github.com/openjournals/joss-reviews/issues/10089#issuecomment-4128491798.
I will fix @Ruslan-Kurmashev 's note about the mistake in qdiv.__version__ in the next release.
