+++ docs/source/qdiv.rst
@@ -1,20 +1,27 @@
 qdiv package
 ============
 
-Main class
-----------
+Overview
+--------
+The `qdiv` package provides tools for microbial diversity analysis built
+around the Hill number framework. Functionality is organized into
+specialized subpackages for diversity, statistics, plotting, modeling,
+and sequence handling, unified by the `MicrobiomeData` container class.
+
+Core subpackages
+----------------
+.. toctree::
+   :maxdepth: 2
+
+   qdiv.diversity
+   qdiv.stats
+   qdiv.plot
+   qdiv.model
+   qdiv.sequences
+
+Main data container
+-------------------
 .. autoclass:: qdiv.MicrobiomeData
    :members:
    :undoc-members:
    :show-inheritance:
-
-Subpackages
------------
-.. toctree::
-   :maxdepth: 4
-
-   qdiv.sequences
-   qdiv.stats
-   qdiv.diversity
-   qdiv.plot
-   qdiv.model
