### 2026-04-27 #10089 @omvatten
@rich2355, I have now released qdiv v4.0.3, which also fixes the version string issue noted in a previous post. I believe all reviewer comments are now fully addressed and responded to above. Please let me know if any further action is required from my side. Thanks again for the constructive review process.
