### 2026-04-27 #10089 @rich2355
> [@rich2355](https://github.com/rich2355), I have now released qdiv v4.0.3, which also fixes the version string issue noted in a previous post. I believe all reviewer comments are now fully addressed and responded to above. Please let me know if any further action is required from my side. Thanks again for the constructive review process.

Hi @Shaun-Jiang @Ruslan-Kurmashev, I just need your final confirmation on whether you think the author has addressed all your comments.

Once I got all of your responses I will move on to the next step. Thanks!
