### 2026-05-01 #10089 @rich2355
Hi @omvatten, please fix the MISSING and INVALID DOIs. In addition, for the section of SKIP DOIs, please add them back if they actually have a DOI (if not, feel free to ignore them).

Please also archive the project into repos like Zonodo, and let me know the link of that archive and the version number.

Thanks!
