+++ paper.bib
@@ -76,6 +76,7 @@
   author = {Jari Oksanen and Gavin L. Simpson and F. Guillaume Blanchet and Roeland Kindt and Pierre Legendre and Peter R. Minchin and R.B. O'Hara and Peter Solymos and M. Henry H. Stevens and Eduard Szoecs and Helene Wagner and Matt Barbour and Michael Bedward and Ben Bolker and Daniel Borcard and Tuomas Borman and Gustavo Carvalho and Michael Chirico and Miquel {De Caceres} and Sebastien Durand and Heloisa Beatriz Antoniazi Evangelista and Rich FitzJohn and Michael Friendly and Brendan Furneaux and Geoffrey Hannigan and Mark O. Hill and Leo Lahti and Cameron Martino and Dan McGlinn and Marie-Helene Ouellette and Eduardo {Ribeiro Cunha} and Tyler Smith and Adrian Stier and Cajo J.F. {Ter Braak} and James Weedon},
   year = {2026},
   note = {R package version 2.8-0},
+  doi = {10.32614/CRAN.package.vegan},
   url = {https://vegandevs.github.io/vegan/},
 }
 
@@ -97,6 +98,7 @@
   year = {2017},
   note = {Microbiome package version 1.31.2},
   organization = {Bioconductor},
+  doi = {10.18129/B9.bioc.microbiome},
   url = {http://microbiome.github.com/microbiome}
 }
 
@@ -114,7 +116,7 @@
   title = {hilldiv: an R package for the integral analysis of diversity based on Hill numbers},
   journal = {bioRxiv},
   year = {2019},
-  doi = {10.1101/545665v1},
+  doi = {10.1101/545665},
   url = {https://www.biorxiv.org/content/10.1101/545665v1}
 }
 
@@ -142,6 +144,7 @@
   author = {Hsieh, T. C. and Ma, K. H. and Chao, A.},
   year = {2024},
   note = {R package version 3.0.0},
+  doi = {10.32614/CRAN.package.iNEXT},
   url = {https://github.com/JohnsonHsieh/iNEXT}
 }
 
@@ -160,7 +163,7 @@
   title = {Scikit-bio: a fundamental Python library for biological omic data analysis},
   journal = {Nature Methods},
   year = {2025},
-  doi = {10.1038/s41592-025-02981-z},
+  doi = {10.1038/s41592-025-02981-z}
 }
 
 @article{huerta2016,
