### 2026-05-03 #10089 @omvatten
@rich2355:
Ok, I have revised the DOIs in paper.bib and I think the missing and invalid are ok now. I also managed to find a few of the missing ones.
I have put qdiv release v4.0.3 on Zenodo. Here is the DOI: https://doi.org/10.5281/zenodo.20004622
